requirements.txt
Flask==3.0.0
PyPDF2==3.0.1
Werkzeug==3.0.1
langchain==0.0.353
pypdf==3.17.4
python-dotenv==1.0.0
openai==1.6.1
chromadb==0.4.21
tiktoken==0.5.2